long double -> double

+++ includes/fractol.h
@@ -6,7 +6,7 @@
 /*   By: susumuyagi <susumuyagi@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/04 15:20:13 by susumuyagi        #+#    #+#             */
-/*   Updated: 2023/11/08 15:28:10 by susumuyagi       ###   ########.fr       */
+/*   Updated: 2023/11/08 15:39:23 by susumuyagi       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,13 +32,13 @@ struct					s_vars
 	void				*mlx;
 	void				*win;
 
-	long double			scale;
-	long double			dx;
-	long double			dy;
+	double				scale;
+	double				dx;
+	double				dy;
 
 	int					is_down;
-	long double			down_x;
-	long double			down_y;
+	double				down_x;
+	double				down_y;
 	int					keycode;
 
 	t_data				img;
@@ -49,8 +49,8 @@ struct					s_vars
 	int					color;
 	int					(*model)(int, int, t_vars *);
 	int					is_julia;
-	long double			c[2];
-	long double			z[WINDOW_WIDTH * WINDOW_HEIGHT * 2];
+	double				c[2];
+	double				z[WINDOW_WIDTH * WINDOW_HEIGHT * 2];
 	int					count[WINDOW_WIDTH * WINDOW_HEIGHT];
 };
