+++ src/mandelbrot.c
@@ -6,7 +6,7 @@
 /*   By: susumuyagi <susumuyagi@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/11 14:16:40 by susumuyagi        #+#    #+#             */
-/*   Updated: 2023/11/08 15:26:38 by susumuyagi       ###   ########.fr       */
+/*   Updated: 2023/11/08 15:39:47 by susumuyagi       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,10 +15,10 @@
 
 int	mandelbrot(int w, int h, t_vars *vars)
 {
-	int			count;
-	long double	c[2];
-	long double	*z;
-	long double	tmp;
+	int		count;
+	double	c[2];
+	double	*z;
+	double	tmp;
 
 	to_z(w, h, c, vars);
 	count = 0;
